# Optimize Array Slicing Performance 