--- a/docs/release-notes/.FSharp.Core/10.0.100.md
+++ b/docs/release-notes/.FSharp.Core/10.0.100.md
@@ -7,5 +7,6 @@
 ### Changed
 
 * Random functions support for zero element chosen/sampled ([PR #18568](https://github.com/dotnet/fsharp/pull/18568))
+* Optimize array slicing performance. ([PR #18778](https://github.com/dotnet/fsharp/pull/18778))
 
 ### Breaking Changes