--- a/docs/release-notes/.FSharp.Core/10.0.100.md
+++ b/docs/release-notes/.FSharp.Core/10.0.100.md
@@ -10,3 +10,5 @@
 * Optimize array slicing performance. ([PR #18778](https://github.com/dotnet/fsharp/pull/18778))
 
 ### Breaking Changes
+
+* Empty 1D array slices return `[||]` instead of allocating a new array. ([PR #18778](https://github.com/dotnet/fsharp/pull/18778))