--- a/docs/release-notes/.FSharp.Core/10.0.100.md
+++ b/docs/release-notes/.FSharp.Core/10.0.100.md
@@ -11,4 +11,4 @@
 
 ### Breaking Changes
 
-* Empty 1D array slices return `[||]` instead of allocating a new array. ([PR #18778](https://github.com/dotnet/fsharp/pull/18778))
+* 1D array slicing now returns an empty array singleton instead of allocating a new array when the result is empty. ([PR #18778](https://github.com/dotnet/fsharp/pull/18778))