--- a/tests/FSharp.Core.UnitTests/FSharp.Core/Microsoft.FSharp.Collections/ArrayModule.fs
+++ b/tests/FSharp.Core.UnitTests/FSharp.Core/Microsoft.FSharp.Collections/ArrayModule.fs
@@ -1963,6 +1963,17 @@ type ArrayModule() =
 
 
     [<Fact>]
+    member _.``Slicing always creates new arrays`` () =
+        let arr = [|1;2;3;4;5;6|]
+
+        Assert.False(Object.ReferenceEquals(arr[1..3], arr[1..3]))
+        Assert.False(Object.ReferenceEquals(arr[0..], arr[0..]))
+        Assert.False(Object.ReferenceEquals(arr[..^1], arr[..^1]))
+        
+        Assert.AreEqual(arr[1..-1].Length, 0)
+        Assert.False(Object.ReferenceEquals(arr[1..-1], arr[1..-1]))
+
+    [<Fact>]
     member _.RandomShuffle() =
         let arr = [| 1..20 |]