## Description

Delegate array slicing to `System.Array.Copy`.

Fixes #18644

## Benchmark Results:
The MicroPerf benchmarks that are committed aren't launching on my installed version of .NET (I don't have .NET 10 Preview 7 access)

Using benchmarks from #18644:

```fsharp
[<MemoryDiagnoser; ShortRunJob>]
type Benchmarks () =
    let b = [| for _ in 1 .. 100_000 -> byte DateTime.Now.Ticks |]

    [<Benchmark>]
    member _.x () =
        b.[ 20 .. 4999 ]

    [<Benchmark>]
    member f.xx () =
        let out = GC.AllocateUninitializedArray<byte> (4980, false)
        Array.Copy (b, 20, out, 0, 4980)
        out

    [<Benchmark>]
    member gg.xxx () =
        b.AsSpan(20, 4980).ToArray ()
```

Before:
```
BenchmarkDotNet v0.15.2, Windows 11 (10.0.26100.4652/24H2/2024Update/HudsonValley)
AMD Ryzen 7 5800X, 1 CPU, 16 logical and 8 physical cores
.NET SDK 10.0.100-preview.6.25358.103
  [Host]   : .NET 10.0.0 (10.0.25.35903), X64 RyuJIT AVX2 DEBUG
  ShortRun : .NET 10.0.0 (10.0.25.35903), X64 RyuJIT AVX2

Job=ShortRun  IterationCount=3  LaunchCount=1
WarmupCount=3

| Method | Mean       | Error     | StdDev   | Gen0   | Allocated |
|------- |-----------:|----------:|---------:|-------:|----------:|
| x      | 2,514.5 ns | 690.39 ns | 37.84 ns | 0.2975 |   4.89 KB |
| xx     |   224.0 ns | 158.36 ns |  8.68 ns | 0.2983 |   4.89 KB |
| xxx    |   210.8 ns |  88.84 ns |  4.87 ns | 0.2992 |   4.89 KB |
```

After:
```
BenchmarkDotNet v0.15.2, Windows 11 (10.0.26100.4652/24H2/2024Update/HudsonValley)
AMD Ryzen 7 5800X, 1 CPU, 16 logical and 8 physical cores
.NET SDK 10.0.100-preview.6.25358.103
  [Host]   : .NET 10.0.0 (10.0.25.35903), X64 RyuJIT AVX2 DEBUG
  ShortRun : .NET 10.0.0 (10.0.25.35903), X64 RyuJIT AVX2

Job=ShortRun  IterationCount=3  LaunchCount=1
WarmupCount=3

| Method | Mean     | Error    | StdDev   | Gen0   | Allocated |
|------- |---------:|---------:|---------:|-------:|----------:|
| x      | 238.1 ns | 108.7 ns |  5.96 ns | 0.2992 |   4.89 KB |
| xx     | 218.4 ns | 210.3 ns | 11.53 ns | 0.2983 |   4.89 KB |
| xxx    | 232.6 ns | 389.8 ns | 21.37 ns | 0.2992 |   4.89 KB |
```

90% performance improvement on my machine. An AVX-512 CPU would be even faster.

## Tests
Existing tests already cover array slicing get/set. This PR adds an additional test to verify an array returned as a slice is unique (different object) as is the existing behavior.

## Checklist

- [x] Test cases added
- [x] Performance benchmarks added in case of performance changes
- [x] Release notes entry updated