@albert-du 2025-07-26
@kerams

Returning `[||]` for a zero length slice might be a breaking change.

Currently, creating two empty slices results in two empty arrays without referential equality. 
```fsharp
let a = [|1;2;3|]

let b = a[1..0]
let c = a[1..0]

printfn $"{b.Length}" // 0
printfn $"{c.Length}" // 0

printfn $"{b = c}" // True
printfn $"{System.Object.ReferenceEquals(b,c)}" // False
```
[sharplab](https://sharplab.io/#v2:DYLgZgzgNALiBOBXAdlAJiA1AH2AUxgAIBDQgXkIG1sBGAbgCY6BmbAXQFgAob/IgI3IlKNAHSiADJy59CAYyHER4qd24AHeAEtkMMMkIASAEQBvfqIAyeZAHMYACwC+xwgHo3hCRu279RszkrG3tnVw8vNS5NHT0DE3MhORd3TwAVJDwfWP8EgGUATwgYPABbUQB5fgArPDkYUQAlPDA8eBs5PABRAEdEYmAIAAp+KDkAShSIgDEBiDwgA=)

By returning `[||]`, all empty slices are represented by the same empty array.

```fsharp
let b = [||]
let c = [||]

printfn $"{b.Length}" // 0
printfn $"{c.Length}" // 0

printfn $"{b = c}" // True
printfn $"{System.Object.ReferenceEquals(b,c)}" // True // this is different
```

[sharplab](https://sharplab.io/#v2:DYLgZgzgNALiBOBXAdlAJiA1AH2AUxgAIAjQgXkIG1tsBdAWACh8iBjcqmhxpgB3gCWyGGGSEAJACIA3sQB0AGTzIA5jAAWAX0mEA9LsIAGPoOGiJM1ouVqtO/UaYmhIsVNkdW2vQYAqSPGczNxkAZQBPCBg8AFs5AHliACs8Vhg5ACU8MDx4ZVY8AFEAR0QAQ2AIAApiKFYASm8Hf0Q8H0INAQhCLsI0ATAcvOEgA==)

This doesn't affect non zero sized slices, so there won't be inconsistent behavior due to array mutation. If there was code dependent on a slice always returning a unique object, regardless of size, (perhaps for a lock) this change would break it.

Returning `[||]` would be more performant, fewer objects for the GC, but I think this change in behavior would need to be noted somewhere.