@brianrourkeboll 2025-07-26
> Returning [||] for a zero length slice might be a breaking change.
>
> Currently, creating two empty slices results in two empty arrays without referential equality.

I think it's an acceptable change. `[||]` itself used to return a new array but no longer does.